+++ glide-core/redis-rs/redis/src/cluster_async/mod.rs
@@ -1162,6 +1162,7 @@ where
             push_sender,
             disconnect_notifier,
             discover_az,
+            connection_timeout: None,
         };
 
         let connections = Self::create_initial_connections(