+++ java/client/src/main/java/glide/managers/ConnectionManager.java
@@ -8,6 +8,7 @@ import connection_request.ConnectionRequestOuterClass.ConnectionRequest;
 import connection_request.ConnectionRequestOuterClass.PubSubChannelsOrPatterns;
 import connection_request.ConnectionRequestOuterClass.PubSubSubscriptions;
 import connection_request.ConnectionRequestOuterClass.TlsMode;
+import glide.api.models.configuration.AdvancedBaseClientConfiguration;
 import glide.api.models.configuration.BaseClientConfiguration;
 import glide.api.models.configuration.GlideClientConfiguration;
 import glide.api.models.configuration.GlideClusterClientConfiguration;
@@ -171,6 +172,30 @@ public class ConnectionManager {
             connectionRequestBuilder.setPubsubSubscriptions(subscriptionsBuilder.build());
         }
 
+        if (configuration.getAdvancedConfiguration() != null) {
+            connectionRequestBuilder =
+                    setupConnectionRequestBuilderAdvancedBaseConfiguration(
+                            connectionRequestBuilder, configuration.getAdvancedConfiguration());
+        }
+
+        return connectionRequestBuilder;
+    }
+
+    /**
+     * Configures the {@link ConnectionRequest.Builder} with settings from the provided {@link
+     * AdvancedBaseClientConfiguration}.
+     *
+     * @param connectionRequestBuilder The builder for the {@link ConnectionRequest}.
+     * @param configuration The advanced configuration settings.
+     * @return The updated {@link ConnectionRequest.Builder}.
+     */
+    private ConnectionRequest.Builder setupConnectionRequestBuilderAdvancedBaseConfiguration(
+            ConnectionRequest.Builder connectionRequestBuilder,
+            AdvancedBaseClientConfiguration configuration) {
+        if (configuration.getConnectionTimeout() != null) {
+            connectionRequestBuilder.setConnectionTimeout(configuration.getConnectionTimeout());
+        }
+
         return connectionRequestBuilder;
     }
 
@@ -197,6 +222,12 @@ public class ConnectionManager {
                         entry.getKey().ordinal(), channelsBuilder.build());
             }
             connectionRequestBuilder.setPubsubSubscriptions(subscriptionsBuilder.build());
+        }
+
+        if (configuration.getAdvancedConfiguration() != null) {
+            connectionRequestBuilder =
+                    setupConnectionRequestBuilderAdvancedBaseConfiguration(
+                            connectionRequestBuilder, configuration.getAdvancedConfiguration());
         }
 
         return connectionRequestBuilder;